[Decomp][Unit] Cohesion split: extract speed/movement-rate methods to UnitSpeed.cpp
